requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.3
backcall==0.1.0
cfgv==3.1.0
decorator==4.4.2
distlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
identify==1.4.14
importlib-metadata==1.6.0
importlib-resources==1.4.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.3.5
numpy==1.18.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pre-commit==2.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyee==7.0.1
Pygments==2.6.1
pylint==2.4.4
pypotrace==0.2
PyYAML==5.3.1
six==1.14.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
virtualenv==20.0.16
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
